libparted: Fail early when detecting nilfs2
The possibly checked LUKS area might be protected with OPAL HW encryption
and not yet fully unlocked, so try to detect nilfs2 at the beginning of the
area first and if that succeeds, detect it at the end of the area. This
prevents having read errors from the locked area.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>